input varification (create new)
