+++ Code/LogicLayer/LogicLayerAPI.py
@@ -2,6 +2,7 @@ from LogicLayer.EmployeeManagerLogic import EmployeeManagerLogic
 from LogicLayer.WorkTripLogic import WorkTripLogic
 from LogicLayer.DestinationManagerLogic import DestinationManagerLogic
 from LogicLayer.AirplaneManagerLogic import AirplaneManagerLogic
+from LogicLayer.IsChecks import IsChecks
 
 
 class LogicLayerAPI:
@@ -10,6 +11,7 @@ class LogicLayerAPI:
         self.work_trip_logic = WorkTripLogic()
         self.destination_logic = DestinationManagerLogic()
         self.airplane_logic = AirplaneManagerLogic()
+        self.checks = IsChecks()
 
     ############################## EmployeeManagerLogic ###############################
     def add_employee(self, employee_type, employee_role, **kwargs):
@@ -227,6 +229,59 @@ class LogicLayerAPI:
         """
         return self.airplane_logic.find_airplane_by_id(airplane_id)
 
+    ############################## Input Varification ###############################
+
+    def is_city(self, City):
+        return self.checks.is_city(City)
+
+    def is_airport(self, Airport):
+        return self.checks.is_airport(Airport)
+
+    def is_country(self, Country):
+        return self.checks.is_country(Country)
+
+    def is_distance(self, Distance):
+        return self.checks.is_distance(Distance)
+
+    def is_travel_time(self, Travel_Time):
+        return self.checks.is_travel_time(Travel_Time)
+
+    def is_contact_name(self, Contact_Name):
+        return self.checks.is_contact_name(Contact_Name)
+
+    def is_contact_phone_number(self, Contact_Phone_Number):
+        return self.checks.is_contact_phone_number(Contact_Phone_Number)
+    
+    def is_name(self, Name):
+        return self.check.is_name(Name)
+
+    def is_current_location(self, Current_Location):
+        return self.check.is_current_location(Current_Location)
+
+    def is_ssn(self, social_security_number):
+        return self.check.is_name(social_security_number)
+
+    def is_type(self, Type):
+        return self.check.is_type(Type)
+
+    def is_manufacturer(self, Manufacturer):
+        return self.check.is_manufacturer(Manufacturer)
+
+    def is_capacity(self, Capacity):
+        return self.check.is_capacity(Capacity)
+
+    def is_address(self, address):
+        return self.check.is_address(address)
+
+    def is_email(self, Email):
+        return self.check.is_email(Email)
+
+    def is_employee_type(self, Employee_Type):
+        return self.check.is_employee_type(Employee_Type)
+
+    def is_employee_role(self, Employee_Role):
+        return self.check.is_employee_role(Employee_Role)
+
     ############################## GeneralUseLogic ###############################
 
     def object_list_to_dict_list(self, object_list):
@@ -234,24 +289,3 @@ class LogicLayerAPI:
 
     def object_to_dict(self, object):
         return self.employee_logic.object_to_dict(object)
-
-    def is_destination(self, City):
-        return self.destination_logic.is_destination(City)
-
-    def is_airport(self, Airport):
-        return self.destination_logic.is_airport(Airport)
-
-    def is_country(self, Country):
-        return self.destination_logic.is_country(Country)
-
-    def is_distance(self, Distance):
-        return self.destination_logic.is_distance(Distance)
-
-    def is_travel_time(self, Travel_Time):
-        return self.destination_logic.is_travel_time(Travel_Time)
-
-    def is_contact_name(self, Contact_Name):
-        return self.destination_logic.is_contact_name(Contact_Name)
-
-    def is_contact_phone_number(self, Contact_Phone_Number):
-        return self.destination_logic.is_contact_phone_number(Contact_Phone_Number)
