+++ Code/UiLayer/EmployeeDataCreateNewUI.py
@@ -244,82 +244,71 @@ class EmployeeDataCreateNewUI:
     def create_new_sequence(self):
         n = 1
         input_check = True
-        value_error = "Value Error string goes here"
         while n < 9:
             if n == 1:
                 self.input_employee_type()
-                if input_check:
-                    data = input("Enter Type: ")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
-                else:
-                    print(value_error)
-                    data = input("Enter Type:")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
+                data = input("Enter Employee Type: ")
+                if data == "q":
+                    print("Goodbye")
+                    exit()
+                try:
+                    self.Logic.is_employee_type(data)
+                    input_check = True
+                except ValueError as e:
+                    print(f"Error: {e}")
+                    input_check = False
 
             elif n == 2:
                 self.input_employee_role()
-                if input_check:
-                    data = input("Enter Role: ")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
-                else:
-                    print(value_error)
-                    data = input("Enter Role:")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
+                data = input("Enter Employee Role: ")
+                if data == "q":
+                    print("Goodbye")
+                    exit()
+                try:
+                    self.Logic.is_employee_role(data)
+                    input_check = True
+                except ValueError as e:
+                    print(f"Error: {e}")
+                    input_check = False
 
             elif n == 3:
                 self.input_name()
-                if input_check:
-                    data = input("Enter Name: ")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
-                else:
-                    print(value_error)
-                    data = input("Enter Name:")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
+                data = input("Enter City: ")
+                if data == "q":
+                    print("Goodbye")
+                    exit()
+                try:
+                    self.Logic.is_destination(data)
+                    input_check = True
+                except ValueError as e:
+                    print(f"Error: {e}")
+                    input_check = False
 
             elif n == 4:
                 self.input_SSN()
-                if input_check:
-                    data = input("Enter SSN: ")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
-                else:
-                    print(value_error)
-                    data = input("Enter SSN (dddddd-dddd):")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
-
-                if data.isnumeric():#------------isSSN()-------------- 
+                data = input("Enter Social Security Number: ")
+                if data == "q":
+                    print("Goodbye")
+                    exit()
+                try:
+                    self.Logic.is_ssn(data)
                     input_check = True
-                else:
+                except ValueError as e:
+                    print(f"Error: {e}")
                     input_check = False
                 
             elif n == 5:
                 self.input_phone()
-                if input_check:
-                    data = input("Enter Phone number: ")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
-                else:
-                    print(value_error)
-                    data = input("Enter Phone number (ddd dddd):")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
+                data = input("Enter Phone: ")
+                if data == "q":
+                    print("Goodbye")
+                    exit()
+                try:
+                    self.Logic.is_contact_phone_number(data)
+                    input_check = True
+                except ValueError as e:
+                    print(f"Error: {e}")
+                    input_check = False
 
                 if data.isnumeric():#------------isPhone()-------------- 
                     input_check = True
@@ -328,49 +317,41 @@ class EmployeeDataCreateNewUI:
 
             elif n == 6:
                 self.input_address()
-                if input_check:
-                    data = input("Enter Address: ")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
-                else:
-                    print(value_error)
-                    data = input("Enter Address:")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
+                data = input("Enter Address: ")
+                if data == "q":
+                    print("Goodbye")
+                    exit()
+                try:
+                    self.Logic.is_address(data)
+                    input_check = True
+                except ValueError as e:
+                    print(f"Error: {e}")
+                    input_check = False
 
             elif n == 7:
                 self.input_email()
-                if input_check:
-                    data = input("Enter Email: ")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
-                else:
-                    print(value_error)
-                    data = input("Enter Email:")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
+                data = input("Enter Email: ")
+                if data == "q":
+                    print("Goodbye")
+                    exit()
+                try:
+                    self.Logic.is_email(data)
+                    input_check = True
+                except ValueError as e:
+                    print(f"Error: {e}")
+                    input_check = False
 
             elif n == 8:
                 self.input_home_phone()
-                if input_check:
-                    data = input("Enter Home Phone: ")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
-                else:
-                    print(value_error)
-                    data = input("Enter Home Phone (ddd dddd):")
-                    if data == "q":
-                        print("Goodbye")
-                        exit()
-                    
-                if data.isnumeric():#------------isPhone()-------------- 
+                data = input("Enter Home Phone: ")
+                if data == "q":
+                    print("Goodbye")
+                    exit()
+                try:
+                    self.Logic.is_contact_phone_number(data)
                     input_check = True
-                else:
+                except ValueError as e:
+                    print(f"Error: {e}")
                     input_check = False
 
             if input_check:
